Tried fixing a dependency problem (check the description!)
org.junit.jupiter cannot be used for defining tests with the @Test tag, or I get a 'Failed to resolve org.junit.vintage:junit-vintage-engine:4.12.12' error.
Using org.junit.* fixes the problem, but only for simpler unit tests.

Adding the junit-vintage-engine dependency causes a NoSuchMethodError; I'm thinking it's an issue with this version of Spring Boot. Tests that involve Spring, and possibly other complex tests won't be implemented for the final version. More basic tests will be added in the next version(s).